third_party/blink/web_tests/platform/generic/http/tests/images/image-with-origin-header-expected.txt

Tests that the order of setting the src of an image element and its crossOrigin attribute doesn't matter

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS srcFirstHasOrigin is true
PASS srcLastHasOrigin is true
PASS successfullyParsed is true

TEST COMPLETE
